refactor(web): home ルートを -components 構造に整理
- HomeComponent を routes/-components/pages/HomePage.tsx に抽出
- components/home/{hero,favorites,lt,blog}.tsx を routes/-components/fragments/ に移動 (PascalCase 化)
- routes/index.tsx を 60 → 6 行に圧縮 (createFileRoute + HomePage 参照だけ)
- 全コンポーネントをアロー関数 named const 表記に統一
- ThreeBackground は汎用 UI として components/backgrounds/ に温存
